Adding works for two sections
There was an issue where the number of times they added a given record
type was the number of times that record was duplicated... first it
would create one; second it would create two... tenth it would create
ten... yeah.  Had to unbind the "click" action at the end, that did the
trick.